Finished hooks, split vector_deskew, reordered functions in vector_utils, revised some documentation
