+++ vision/coordinate_lengths.py
@@ -1,3 +1,5 @@
+"""Functions for calculating coordinate degree lengths"""
+
 import numpy as np
 
 
@@ -10,6 +12,11 @@ def latitude_length(latitude: float) -> float:
     latitude : float
         The latitude in degrees
 
+    Returns
+    -------
+    latitude_length
+        The length of a degree of latitude in meters at the given latitude
+
     References
     ----------
     https://en.wikipedia.org/wiki/Geographic_coordinate_system#Length_of_a_degree
@@ -18,8 +25,12 @@ def latitude_length(latitude: float) -> float:
     # Convert to radians for trig functions
     latitude = np.deg2rad(latitude)
 
-    distance: float = 111132.92 - 559.82 * np.cos(2 * latitude) + 1.175 * np.cos(4 * latitude)\
+    distance: float = (
+        111132.92
+        - 559.82 * np.cos(2 * latitude)
+        + 1.175 * np.cos(4 * latitude)
         - 0.0023 * np.cos(6 * latitude)
+    )
 
     return distance
 
@@ -35,8 +46,8 @@ def longitude_length(latitude: float) -> float:
 
     Returns
     -------
-
-
+    longitude_length
+        The length of a degree of longitude in meters at the given latitude
     References
     ----------
     https://en.wikipedia.org/wiki/Geographic_coordinate_system#Length_of_a_degree
@@ -45,7 +56,8 @@ def longitude_length(latitude: float) -> float:
     # Convert degrees to radians for trig functions
     latitude = np.deg2rad(latitude)
 
-    distance: float = 111412.84 * np.cos(latitude) - 93.5 * np.cos(3 * latitude) \
-        + 0.118 * np.cos(5 * latitude)
+    distance: float = (
+        111412.84 * np.cos(latitude) - 93.5 * np.cos(3 * latitude) + 0.118 * np.cos(5 * latitude)
+    )
 
     return distance
